Fix #1582: support reading PDF files larger than 2 GB
File positions were tracked as int in RandomAccessFileOrArray,
PRTokeniser, PRStream and the PdfReader xref table, so offsets beyond
Integer.MAX_VALUE overflowed and reading a >2 GB PDF failed with
'Rebuild failed: Position out of bounds; Original message: PDF startxref
not found' (or scanned the file for a very long time before failing).

Switch the read path to 64-bit offsets:
- RandomAccessFileOrArray: length(), getFilePointer(), startOffset and
  skip() are now long-based; seek(long) is the real implementation.
- PRTokeniser: seek/getFilePointer/length/getStartxref use long; new
  longValue() for parsing 10-digit offsets.
- PdfReader: xref table is long[], lastXref/eofPos/fileLength are long,
  xref-stream type-1 offsets read as long, objStmToOffset maps to Long.
- PRStream: stream offset is long.
- PdfNumber: new longValue() accessor.
- PdfWriter: prevxref is long so incremental updates of large files
  write a correct /Prev.
- MappedRandomAccessFile was already long-based (LongMappedByteBuffer).

Font-parsing call sites (TrueTypeFont, Type1Font, EnumerateTTC, CFFFont)
updated for the new signatures; font files remain int-sized.

The regression test builds a sparse PDF whose objects live above the
2 GB boundary (a few KB of real data), so it runs in under a second; a
30s timeout guards against regressions, which would manifest as
long scans rather than fast failures.

+++ openpdf-core/src/main/java/org/openpdf/text/pdf/PRStream.java
@@ -61,7 +61,7 @@ import java.util.zip.DeflaterOutputStream;
 public class PRStream extends PdfStream {
 
     protected PdfReader reader;
-    protected int offset;
+    protected long offset;
     protected int length;
 
     //added by ujihara for decryption
@@ -90,7 +90,7 @@ public class PRStream extends PdfStream {
         this.reader = reader;
     }
 
-    public PRStream(PdfReader reader, int offset) {
+    public PRStream(PdfReader reader, long offset) {
         this.reader = reader;
         this.offset = offset;
     }
@@ -182,7 +182,7 @@ public class PRStream extends PdfStream {
         setData(data, true);
     }
 
-    public int getOffset() {
+    public long getOffset() {
         return offset;
     }
