+++ openpdf-core/src/main/java/org/openpdf/text/pdf/TrueTypeFont.java
@@ -516,7 +516,7 @@ class TrueTypeFont extends BaseFont {
             int length = rf.readUnsignedShort();
             int offset = rf.readUnsignedShort();
             if (nameID == id) {
-                int pos = rf.getFilePointer();
+                long pos = rf.getFilePointer();
                 rf.seek(table_location[0] + startOfStorage + offset);
                 String name;
                 if (platformID == 0 || platformID == 3 || (platformID == 2 && platformEncodingID == 1)) {
@@ -559,7 +559,7 @@ class TrueTypeFont extends BaseFont {
             int nameID = rf.readUnsignedShort();
             int length = rf.readUnsignedShort();
             int offset = rf.readUnsignedShort();
-            int pos = rf.getFilePointer();
+            long pos = rf.getFilePointer();
             rf.seek(table_location[0] + startOfStorage + offset);
             String name;
             if (platformID == 0 || platformID == 3 || (platformID == 2 && platformEncodingID == 1)) {
@@ -1257,7 +1257,7 @@ class TrueTypeFont extends BaseFont {
         try {
             rf2 = new RandomAccessFileOrArray(rf);
             rf2.reOpen();
-            byte[] b = new byte[rf2.length()];
+            byte[] b = new byte[(int) rf2.length()];
             rf2.readFully(b);
             return b;
         } finally {
